lint and rearrange
